organize components between common and bootstrap
